fixing warnings refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@25042 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSVehicle.cpp
@@ -814,7 +814,7 @@ Position
 MSVehicle::validatePosition(Position result, double offset) const {
     int furtherIndex = 0;
     while ((ISNAN(result.x()) || result == Position::INVALID)) {
-        if (furtherIndex >= myFurtherLanes.size()) {
+        if (furtherIndex >= (int)myFurtherLanes.size()) {
             //WRITE_WARNING("Could not compute position for vehicle '" + getID() + "', time=" + time2string(MSNet::getInstance()->getCurrentTimeStep()) + ".");
             break;
         }
